+++ src/day06/blog/internal/database/schema/001_articles.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 
 CREATE TABLE articles (
-    id bigint NOT NULL PRIMARY KEY,
+    id SERIAL NOT NULL PRIMARY KEY,
     title TEXT NOT NULL,
     content TEXT NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
